installer fixes

+++ install_scripts/adafruit-pitft-setup.sh
@@ -798,12 +798,12 @@ info PITFT "Success!"
 echo
 echo "Settings take effect on next boot."
 echo
-#echo -n "REBOOT NOW? [y/N] "
-#read
-#if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then
-#        echo "Exiting without reboot."
-#        exit 0
-#fi
+echo -n "REBOOT NOW? [y/N] "
+read
+if [[ ! "$REPLY" =~ ^(yes|y|Y)$ ]]; then
+        echo "Exiting without reboot."
+        exit 0
+fi
 
 echo "Reboot started..."
 reboot
